Update to API, Operation, and ape-lib
The API has been updated to send a proper response to a request.
Responses are json encoded with a header having a content-type of
application/json.

The abstract Operation class has been updated to build up build
up the start of a response array object based on a condition.
It still needs to handle several other conditions, including
validation.

The concrete Name class (as a proof of concept) has been updated
to interact with new parent behavior, in addition to returning
the completely built up response object to the API.

The api-lib javascript library has been updated to common
ajax functionality that can be used to interact with the API.
Additionally, the name function has been updated as a PoC to
send the API a "name" operation with a collection of parameters.
In this case, that collection results to one parameter called
name. If additional parameters are sent, the appropriate error
response (Invalid number of parameters) is detected and sent back
(with the "status" prop being equal to "false" and with the
"message" prop being equal to the appropriate error message).