Update to API+Operation, Create ExamDetails
The API has been updated to determining the success and message
response, as opposed to having the Operation classes deal with
that. The API will also pass the accountID as defined by the
$_SESSION variable.

The abstract Operation class has been updated to remove the
"mixed" psuedo-type from method signatures as a type-hint.
I was running into an error when invoking said functions saying,
"Argument must be an instance of mixed, string passed"
Upon investigating the PHP 7 documentation, it appears that
"mixed" is a keyword but not a language keyword, and only
language keywords can be specified for typehints.

I've also added a case to Operation's execute method where if the
$args is an array, then invoke call_user_func_array (else, invoke
call_user_func).

A concrete ExamDetails class has been made that invokes its parent
registration methods. A corresponding javascript function was made
to invoke this operation, which is tested and working.