Update API and Operation
Updated how the API and Operation coordinate with respect to
error messages. If the abstract Operation class ever throws an
exception, that exception will be exposed to the API. The API
will then take that Exception and add the message to the message
object that is sent back to the user.

The API can be updated to send back less verbose Exception errors
in the future.