Added dataTables vendor
Due to trouble with the bower package of datatables-bs, a
generator was used to combine plugins(js)/styles(css) together
into minified files. Those are being included for now - if there
is enough time in the future, more time can be spent refactoring
bower and the base.twig.html to get the bower vendor files to
work properly.