Added all exam states to Exams page
Added tabs for all of the exam states to the Exams page. The
exam-search component is being reused. Some alterations have been
made to the exam-search component for the following reasoning:

If you include the same component more than once on the same page,
but try to have javascript variables have different values based
on a parm passed to each respective component, you will run into
unintended behavior because javascript variables will override
one another.

The solution is to load the content for component as you need it,
for example, on an onClick event.

The problem with that approach by itself is that if you try to
change the visibility of twig markup that has already been
included in the DOM, there does not exist a way to engage in
variable substitution/interpolation. The interpolation happens
when the page is first rendered by the php backend.

In order to account for this, a new php file ./api/component.php
was added. Similar to modal.php , it takes a componentName and
a collection of parameters. That ./api/component.php file will
then engage in checks to see if the component exists, and if it
does, will invoke the renderPage(...) function using the component
and passed parameters.

The advantage of this is that, on some onclick event, we can
invoke the jQuery load function against ./api.component.php
to update an area of the DOM using a freshly rendered component.
You can now include multiple components on the same page using
this strategy.

./api/component.php can use some improvements, but it is a quick
solution that seems to work and will allow for us to solve similar
problems using this strategy.